+++ coroutines/src/commonMain/kotlin/com/inkapplications/coroutines/ongoing/Collectors.kt
@@ -1,5 +1,6 @@
 package com.inkapplications.coroutines.ongoing
 
+import com.inkapplications.coroutines.whenTrue
 import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.flow.collectLatest
 
@@ -25,5 +26,15 @@ suspend inline fun <T> OngoingFlow<T>.collectLatest(
     crossinline observer: suspend (T) -> Unit
 ): Nothing {
     asFlow().collectLatest { observer(it) }
+    throw UnexpectedEndOfFlow()
+}
+
+/**
+ * @see whenTrue
+ */
+suspend inline fun OngoingFlow<Boolean>.whenTrue(
+    crossinline observer: suspend () -> Unit
+): Nothing {
+    asFlow().whenTrue { observer() }
     throw UnexpectedEndOfFlow()
 }